+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [20171020]
+
+  - Add Android autofocus, refs #2
+  - Fix Android rotation, refs #3
+  - Add animated demo gif
+
 ## [20171019]
 
   - Add ZBar Android support, refs #1